\name{MTurkRGUI-package}
\alias{MTurkRGUI-package}
\alias{MTurkRGUI}
\docType{package}
\title{A graphical user interface (GUI) for the MTurkR package.}
\description{
This package provides a graphical user interface for the MTurkR package. The functionality was previously part of MTurkR, but has been factored out into a separate package to simply development. Users should see the MTurkR documentation for }
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{http://www.mturk.com}{Amazon Mechanical Turk}

\href{http://aws.amazon.com/documentation/mturk/}{Amazon Mechanical Turk API Documentation}

\href{http://www.github.com/leeper/MTurkR}{MTurkR Wiki}
}
\keyword{ package }
%\seealso{}
%\examples{}
